// face-check-ui/src/app/services/models/remote-verification-admin-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface RemoteVerificationAdminResponse {
  companyId?: number;
  companyName?: string;
  createdAt?: string;
  formattedVerificationTime?: string;
  isMissed?: boolean;
  isMissedMessage?: string;
  isSuccessful?: boolean;
  latitude?: number;
  locationAddress?: string;
  longitude?: number;
  message?: string;
  photoUrl?: string;
  status?: 'PENDING' | 'COMPLETED' | 'MISSED';
  verificationId?: number;
  verificationTime?: string;
  workerEmail?: string;
  workerFirstName?: string;
  workerId?: number;
  workerLastName?: string;
  workerPhone?: string;
}