// face-check-ui/src/app/services/models/page-response-remote-verification-admin-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { RemoteVerificationAdminResponse } from '../models/remote-verification-admin-response';
export interface PageResponseRemoteVerificationAdminResponse {
  content?: Array<RemoteVerificationAdminResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElement?: number;
  totalPages?: number;
}